large refactor of content and README.md
so that the table had clearer info and fewer 'N/A' cells, and also that the folders were named according to the `identifier` of each kind of reference.

term, preterm, etc references for the same 'type' or reference eg UK-WHO, ahve been grouped as it is likely they would be used together.